ci: add lerna CLIs to correctly publish the packages.
Adds to the workflow the commands which will firstly, build every
package inside the "packages" folder.
Afterward, the version command will be applied, relating the tags
of the version to the tag from the repository.  As its final command,
sets the publishment of the packages of the monorepo will be
compared to the git last tag.